+++ app/src/main/java/com/bernaferrari/changedetection/detailsPdf/PdfFragment.kt
@@ -1,10 +1,13 @@
-package com.bernaferrari.changedetection.screenDiffImage
+package com.bernaferrari.changedetection.detailsPdf
 
 import android.arch.lifecycle.Observer
 import android.arch.lifecycle.ViewModelProviders
 import android.content.Intent
+import android.graphics.Bitmap
 import android.graphics.Point
+import android.graphics.pdf.PdfRenderer
 import android.os.Bundle
+import android.os.ParcelFileDescriptor
 import android.support.transition.AutoTransition
 import android.support.transition.TransitionManager
 import android.support.v4.app.Fragment
@@ -26,10 +29,9 @@ import com.bernaferrari.changedetection.Application
 import com.bernaferrari.changedetection.R
 import com.bernaferrari.changedetection.ViewModelFactory
 import com.bernaferrari.changedetection.data.Snap
+import com.bernaferrari.changedetection.detailsText.TextFragment
 import com.bernaferrari.changedetection.extensions.*
 import com.bernaferrari.changedetection.groupie.RowItem
-import com.bernaferrari.changedetection.screenDiffText.TextFragment
-import com.bernaferrari.changedetection.util.GlideApp
 import com.davemorrissey.labs.subscaleview.ImageSource
 import com.xwray.groupie.GroupAdapter
 import com.xwray.groupie.Section
@@ -50,42 +52,81 @@ import kotlin.reflect.KProperty
 // If you are interested, download it and scroll the initial list of items for a large button named "Dicionário":
 // https://play.google.com/store/apps/details?id=com.biblialibras.android
 //
-class ImageFragment : Fragment(),
+// This is also adapted from Lottie, which currently provides one of the best open source
+// sample app for a library I've ever seen: https://github.com/airbnb/lottie-android
+//
+class PdfFragment : Fragment(),
     DiscreteScrollView.OnItemChangedListener<RecyclerView.ViewHolder> {
 
-    lateinit var model: ImageViewModel
-    lateinit var adapter: ImageAdapter
-
-    private var previousAdapterPosition = 0
-    private val items = mutableListOf<RowItem>()
-    private val section = Section()
+    lateinit var model: PdfViewModel
+    lateinit var adapter: PdfAdapter
 
     private val transition = AutoTransition().apply { duration = 175 }
     private val uiState = UiState { updateUiFromState() }
 
-    private val groupAdapter = GroupAdapter<com.xwray.groupie.ViewHolder>()
+    private var mPdfRenderer: PdfRenderer? = null
+    private var mCurrentPage: PdfRenderer.Page? = null
+
+    // this variable is necessary since the CurrentPage index changes on updateFileDescriptor, and
+    // this value is needed on updateUiFromState when uiState.highQuality is toggled, so it keeps the same page opened.
+    private var currentIndex = 0
 
     // this variable is necessary since onCurrentItemChanged might be triggered when RecyclerView is shown/hidden
     // (visibility button). This way, the app never has to refresh the file
-    private var currentFileId = ""
+    private var currentFileName = ""
+
+    // this variable is used to track the last position, so the app knows the direction the user is going
+    // and can provide the correct padding. Also, it is used for items selection/deselection.
+    private var previousAdapterPosition = 0
+
+    private val items = mutableListOf<RowItem>()
+    private val section = Section()
+
+    private fun updateUiFromState() {
+        beginDelayedTransition()
+
+        carouselRecycler.isVisible = uiState.carousel
+        controlBar.isVisible = uiState.controlBar
+
+        if (uiState.highQuality != highQualityToggle.isActivated) {
+            highQualityToggle.isActivated = uiState.highQuality
+
+            val item = adapter.getItemFromAdapter(previousAdapterPosition) ?: return
+            updateFileDescriptor(item.snapId)
+            showPage(currentIndex)
+        }
+    }
+
+    override fun onStop() {
+        // this is necessary to avoid java.lang.IllegalStateException: Already closed
+        // if (mCurrentPage != null) isn't working when sharing
+        try {
+            mCurrentPage?.close()
+        } catch (e: IllegalStateException) {
+
+        }
+
+        try {
+            mPdfRenderer?.close()
+        } catch (e: IllegalStateException) {
+
+        }
+
+        super.onStop()
+    }
+
+    private fun beginDelayedTransition() =
+        TransitionManager.beginDelayedTransition(container, transition)
 
     override fun onCurrentItemChanged(viewHolder: RecyclerView.ViewHolder?, adapterPosition: Int) {
 
+        // set the photo_view with current file
         val item = adapter.getItemFromAdapter(adapterPosition) ?: return
         titlecontent.text = item.timestamp.convertTimestampToDate()
 
-        if (item.snapId == currentFileId) return
-        currentFileId = item.snapId
-
-        val newFile = File(Application.instance.filesDir, item.snapId)
-        val contentUri =
-            FileProvider.getUriForFile(
-                Application.instance,
-                "com.bernaferrari.changedetection.files",
-                newFile
-            )
-
-        photo_view.setImage(ImageSource.uri(contentUri))
+        if (updateFileDescriptor(item.snapId)) {
+            showPage(0)
+        }
 
         if (items.isEmpty()) {
             return
@@ -94,31 +135,35 @@ class ImageFragment : Fragment(),
         selectItem(adapterPosition)
     }
 
+    private val groupAdapter = GroupAdapter<com.xwray.groupie.ViewHolder>()
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View = inflater.inflate(R.layout.diff_image_fragment, container, false)
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
 
         model = obtainViewModel(requireActivity())
 
-        closecontent.setOnClickListener {
+        view.closecontent.setOnClickListener {
             view.let { Navigation.findNavController(it).navigateUp() }
         }
 
-        next_previous_bar.isVisible = false
-        highQualityToggle.isVisible = false
-        controlBar.isVisible = false
-
         view.menucontent.setOnClickListener { view.drawer.openDrawer(GravityCompat.END) }
+
+        highQualityToggle.setOnClickListener { uiState.highQuality++ }
+
+        mButtonPrevious.setOnClickListener { showPage(currentIndex - 1) }
+
+        mButtonNext.setOnClickListener { showPage(currentIndex + 1) }
 
         shareToggle.setOnClickListener {
             val item =
                 adapter.getItemFromAdapter(previousAdapterPosition) ?: return@setOnClickListener
             shareItem(item)
         }
-
 
         // this is needed. If visibility is off and the fragment is reopened,
         // drawable will keep the drawable from                                                                                                                                                                                                                                                      last state (off) even thought it should be on.
@@ -133,7 +178,7 @@ class ImageFragment : Fragment(),
             uiState.visibility++
 
             uiState.carousel = uiState.visibility
-            uiState.controlBar = false
+            uiState.controlBar = uiState.visibility
 
             // set and run the correct animation
             if (uiState.visibility) {
@@ -146,16 +191,12 @@ class ImageFragment : Fragment(),
         val recyclerListener = object :
             TextFragment.Companion.RecyclerViewItemListener {
             override fun onClickListener(item: RecyclerView.ViewHolder) {
-                carouselRecycler.smoothScrollToPosition((item as ImageViewHolder).itemPosition)
+                carouselRecycler.smoothScrollToPosition((item as PdfViewHolder).itemPosition)
             }
 
             override fun onLongClickListener(item: RecyclerView.ViewHolder) {
-                removeItemDialog((item as ImageViewHolder).currentSnap?.snapId ?: "")
+                removeItemDialog((item as PdfViewHolder).currentSnap?.snapId ?: "")
             }
-        }
-
-        menucontent.setOnClickListener {
-            view.drawer.openDrawer(GravityCompat.END)
         }
 
         val windowDimensions = Point()
@@ -163,21 +204,47 @@ class ImageFragment : Fragment(),
         val itemWidth = Math.round(Math.min(windowDimensions.y, windowDimensions.x) * 0.5f)
         val itemHeight = (120 * 0.7).toInt().toDp(view.resources)
 
-        adapter = ImageAdapter(
+        adapter = PdfAdapter(
             recyclerListener,
             itemHeight,
             itemWidth,
-            GlideApp.with(this)
+            requireContext()
         )
 
+        carouselRecycler.also {
+            it.adapter = adapter
+
+            it.setSlideOnFling(true)
+            it.setSlideOnFlingThreshold(4000)
+            it.setItemTransitionTimeMillis(150)
+            it.setItemTransformer(
+                ScaleTransformer.Builder()
+                    .setMinScale(0.8f)
+                    .build()
+            )
+
+            it.addOnItemChangedListener(this@PdfFragment)
+        }
+
+        drawerRecycler.also {
+            it.layoutManager = LinearLayoutManager(this.context)
+            it.adapter = groupAdapter
+            it.addItemDecoration(
+                DividerItemDecoration(
+                    this.context,
+                    DividerItemDecoration.VERTICAL
+                )
+            )
+        }
+
         val siteId = arguments?.getString("SITEID") ?: ""
-        model.getAllSnapsPagedForId(siteId, "image%").observe(this, Observer(adapter::submitList))
+        model.getAllSnapsPagedForId(siteId, "%pdf").observe(this, Observer(adapter::submitList))
 
         // this is needed since getSnapsFiltered retrieves a liveData from Room to be observed
         launch {
-            val liveData = model.getSnapsFiltered(siteId, "image%")
+            val liveData = model.getSnapsFiltered(siteId, "%pdf")
             launch(UI) {
-                liveData.observe(this@ImageFragment, Observer {
+                liveData.observe(this@PdfFragment, Observer {
                     val isItemsEmpty = items.isEmpty()
                     items.clear()
 
@@ -201,32 +268,6 @@ class ImageFragment : Fragment(),
             }
         }
 
-        carouselRecycler.also {
-            it.adapter = adapter
-
-            it.setSlideOnFling(true)
-            it.setSlideOnFlingThreshold(4000)
-            it.setItemTransitionTimeMillis(150)
-            it.setItemTransformer(
-                ScaleTransformer.Builder()
-                    .setMinScale(0.8f)
-                    .build()
-            )
-
-            it.addOnItemChangedListener(this@ImageFragment)
-        }
-
-        drawerRecycler.also {
-            it.layoutManager = LinearLayoutManager(this.context)
-            it.adapter = groupAdapter
-            it.addItemDecoration(
-                DividerItemDecoration(
-                    this.context,
-                    DividerItemDecoration.VERTICAL
-                )
-            )
-        }
-
         groupAdapter.add(section)
         groupAdapter.setOnItemClickListener { item, _ ->
             if (item is RowItem) {
@@ -242,6 +283,106 @@ class ImageFragment : Fragment(),
             }
             true
         }
+    }
+
+    private fun shareItem(item: Snap) {
+        val file = File(requireContext().cacheDir, "share.pdf")
+        file.createNewFile()
+        file.writeBytes(requireContext().openFileInput(item.snapId).readBytes())
+
+        val contentUri = FileProvider.getUriForFile(
+            requireContext(),
+            "com.bernaferrari.changedetection.files",
+            file
+        )
+
+        val shareIntent = ShareCompat.IntentBuilder.from(activity)
+            .setStream(contentUri)
+            .setType(item.contentType)
+            .intent
+
+        shareIntent.data = contentUri
+        shareIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
+
+        startActivity(shareIntent)
+    }
+
+    private fun removeItemDialog(snapId: String) {
+        MaterialDialog.Builder(requireContext())
+            .title(R.string.remove)
+            .content(R.string.remove_content)
+            .positiveText(R.string.yes)
+            .negativeText(R.string.no)
+            .onPositive { _, _ ->
+                model.removeSnap(snapId)
+            }
+            .show()
+    }
+
+    private fun updateFileDescriptor(fileName: String): Boolean {
+        if (currentFileName == fileName) return false
+        currentFileName = fileName
+
+        val file = File("${Application.instance.filesDir.absolutePath}/$fileName")
+
+        val mFileDescriptor =
+            ParcelFileDescriptor.open(file, ParcelFileDescriptor.MODE_READ_ONLY)
+
+        mPdfRenderer = PdfRenderer(mFileDescriptor)
+        return true
+    }
+
+
+    /**
+     * Shows the specified page of PDF to the screen.
+     *
+     * @param index The page index.
+     */
+    private fun showPage(index: Int) {
+
+        currentIndex = index
+
+        if (mPdfRenderer!!.pageCount <= index) {
+            if (mCurrentPage != null) {
+                updateUi()
+            }
+
+            return
+        }
+
+        // Make sure to close the current page before opening another one.
+        // The exception catches "Already closed"
+        try {
+            mCurrentPage?.close()
+        } catch (e: IllegalStateException) {
+
+        }
+
+        // Use `openPage` to open a specific page in PDF.
+        mCurrentPage = mPdfRenderer!!.openPage(index).also { mCurrentPage ->
+
+            val qualityMultiplier = if (uiState.highQuality) 4 else 2
+
+            // Important: the destination bitmap must be ARGB (not RGB).
+            val bitmap = Bitmap.createBitmap(
+                mCurrentPage.width * qualityMultiplier, mCurrentPage.height * qualityMultiplier,
+                Bitmap.Config.ARGB_8888
+            )
+            // Here, we render the page onto the Bitmap.
+            // To render a portion of the page, use the second and third parameter. Pass nulls to get
+            // the default result.
+            // Pass either RENDER_MODE_FOR_DISPLAY or RENDER_MODE_FOR_PRINT for the last parameter.
+            mCurrentPage.render(
+                bitmap,
+                null,
+                null,
+                PdfRenderer.Page.RENDER_MODE_FOR_DISPLAY
+            )
+            // We are ready to show the Bitmap to user.
+            view?.photo_view?.setImage(ImageSource.bitmap(bitmap))
+        }
+
+        updateUi()
     }
 
     private fun selectItem(adapterPosition: Int) {
@@ -267,59 +408,23 @@ class ImageFragment : Fragment(),
         section.notifyChanged()
     }
 
-    private fun shareItem(item: Snap) {
-        // images/jpeg should become jpeg
+    /**
+     * Updates the state of 2 control buttons in response to the current page index.
+     */
+    private fun updateUi() {
+        val index = mCurrentPage!!.index
+        val pageCount = mPdfRenderer!!.pageCount
+        mButtonPrevious.isEnabled = 0 != index
+        mButtonNext.isEnabled = index + 1 < pageCount
 
-        val fileExtension = item.contentType.split("/").getOrNull(1) ?: ""
-
-        val file = File(requireContext().cacheDir, "share.$fileExtension")
-        file.createNewFile()
-        file.writeBytes(requireContext().openFileInput(item.snapId).readBytes())
-
-        val contentUri = FileProvider.getUriForFile(
-            requireContext(),
-            "com.bernaferrari.changedetection.files",
-            file
-        )
-
-        val shareIntent = ShareCompat.IntentBuilder.from(activity)
-            .setStream(contentUri)
-            .intent
-
-        shareIntent.type = item.contentType
-        shareIntent.data = contentUri
-        shareIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
-
-        startActivity(shareIntent)
+        next_previous_bar.isVisible = mButtonPrevious.isEnabled || mButtonNext.isEnabled
     }
 
-    private fun removeItemDialog(snapId: String) {
-        MaterialDialog.Builder(requireContext())
-            .title(R.string.remove)
-            .content(R.string.remove_content)
-            .positiveText(R.string.yes)
-            .negativeText(R.string.no)
-            .onPositive { _, _ ->
-                model.removeSnap(snapId)
-            }
-            .show()
-    }
-
-    private fun obtainViewModel(activity: FragmentActivity): ImageViewModel {
+    private fun obtainViewModel(activity: FragmentActivity): PdfViewModel {
         // Use a Factory to inject dependencies into the ViewModel
         val factory = ViewModelFactory.getInstance(activity.application)
-        return ViewModelProviders.of(activity, factory).get(ImageViewModel::class.java)
+        return ViewModelProviders.of(activity, factory).get(PdfViewModel::class.java)
     }
-
-    private fun updateUiFromState() {
-        beginDelayedTransition()
-
-        carouselRecycler.isVisible = uiState.carousel
-        controlBar.isVisible = uiState.controlBar
-    }
-
-    private fun beginDelayedTransition() =
-        TransitionManager.beginDelayedTransition(container, transition)
 
     companion object {
         private class UiState(private val callback: () -> Unit) {
